CASSGO-92: add public method to retrieve StatementMetadata and LogField methods
StatementMetadata can be called on a session to get the bind, result,
and pk information for a given query. Previously this wasn't publicly
exposed but is necessary for some implementations of HostSelectionPolicy
like token-aware. This might also be useful for CI tooling or runtime
analysis of queries and the types of columns.

NewLogField* are methods to to return a LogField with name and a specific
type.

Finally, session init was cleaned up to prevent a HostSelectionPolicy
from causing a panic if it tried to make a query during init. The
interface was documented that queries should not be attempted.

Patch by James Hartig for CASSGO-92; reviewed by João Reis for CASSGO-92